+++ planning/MCRPlannerGoalSet.cpp
@@ -925,7 +925,7 @@ void MCRPlannerGoalSet::KNN(const Config& q,int k,vector<int>& neighbors,vector<
 	}
 	*/
       if(d < dmax) {
-	pair<Real,int> idx(d,i);
+	pair<Real,int> idx(d,(int)i);
 	knn.insert(idx);
 	if((int)knn.size() > k)
 	  knn.erase(--knn.end());
